Refactor 0810

+++ public/js/dashboard.js
@@ -1,6 +1,11 @@
 let access_token = localStorage.getItem('access_token');
 let username = localStorage.getItem('username');
+let username_abv = username.split("")[0];
 let user_id = localStorage.getItem('user_id');
+
+/*---Page Style---*/
+$('.greeting').html(`${username_abv}`);
+
 /*---JQuery Tooltip---*/
 $(function () {
     $(document).tooltip();
@@ -15,54 +20,67 @@ fetch(`/api/1.0/dashboard/${user_id}`, {
     }
 })
     .then((res) => res.json())
-    .then((data) => {
-        for (let i = 0; i < data.length; i++) {
-            $('.whiteboard').append(`
-            <div class='wb_block hvr-grow' id='${data[i].wb_id}'>
-                <div class='wb_bookmark'></div>
-                <div class='wb_back'></div>
-                <input class='wb_title'><img class='edit old' src='./img/edit.png'>
-                <div class='close_btn'>X</div>
-            </div>`)
-            // set title/bookmark on each wb
-            $(`#${data[i].wb_id} > .wb_title`).val(`${data[i].title}`)
-            if (data[i].bookmark == 'bookmarked') {
-                $(`#${data[i].wb_id} > .wb_bookmark`).data('bookmark', `${data[i].bookmark}`)
-                $(`#${data[i].wb_id} > .wb_bookmark`).addClass('bookmarked')
+    .then((json_data) => {
+        if (json_data.error) {
+            Swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: `${json_data.error}`,
+                onAfterClose: () => {
+                    location.href = '/';
+                }
+            });
+        } else {
+            let data = json_data.all_wb;
+            for (let i = 0; i < data.length; i++) {
+                $('.whiteboard').append(`
+                <div class='wb_block hvr-grow' id='${data[i].wb_id}'>
+                    <div class='wb_bookmark'></div>
+                    <div class='wb_back'></div>
+                    <input class='wb_title'><img class='edit old' src='./img/edit.png'>
+                    <div class='close_btn'>X</div>
+                </div>`);
+                // set title/bookmark on each wb
+                $(`#${data[i].wb_id} > .wb_title`).val(`${data[i].title}`);
+                if (data[i].bookmark == 'bookmarked') {
+                    $(`#${data[i].wb_id} > .wb_bookmark`).data('bookmark', `${data[i].bookmark}`);
+                    $(`#${data[i].wb_id} > .wb_bookmark`).addClass('bookmarked');
+                }
+                // set "guest" wb to read-only
+                if (data[i].role == 'guest') {
+                    $(`#${data[i].wb_id}`).append(`<img class='readonly' src='../img/lock.png'>`);
+                    $(`#${data[i].wb_id}`).attr('title', `Title editing not authorized`);
+                    $(`#${data[i].wb_id} > .edit, #${data[i].wb_id} > .close_btn`).remove();
+                    $(`#${data[i].wb_id} > .wb_bookmark`).remove();
+                }
             }
-            // set "guest" wb to read-only
-            if (data[i].role == 'guest') {
-                $(`#${data[i].wb_id}`).append(`<img class='readonly' src='../img/lock.png'>`)
-                $(`#${data[i].wb_id}`).attr('title', `Title editing not authorized`);
-                $(`#${data[i].wb_id} > .edit, #${data[i].wb_id} > .close_btn`).remove()
-                $(`#${data[i].wb_id} > .wb_bookmark`).remove()
-            }
+
+            $('.whiteboard').on('mouseover', '.wb_block', function () {
+                $(this).children('.close_btn').show();
+                $(this).children('.edit').css({ 'visibility': 'visible' });
+
+            });
+            $('.whiteboard').on('mouseout', '.wb_block', function () {
+                $(this).children('.close_btn').hide();
+                $(this).children('.edit').css({ 'visibility': 'hidden' });
+            });
+            return;
+
         }
 
-        $('.whiteboard').on('mouseover', '.wb_block', function () {
-            $(this).children('.close_btn').show();
-            $(this).children('.edit').css({ 'visibility': 'visible' })
-
-        })
-        $('.whiteboard').on('mouseout', '.wb_block', function () {
-            $(this).children('.close_btn').hide();
-            $(this).children('.edit').css({ 'visibility': 'hidden' })
-        })
-        return;
-
-    })
+    });
 
 /*---Add whiteboard by add_btn---*/
 // set add_btn bg color 
 $('.add_btn').on('click', function () {
-    $(this).css({ 'background-color': 'white', 'color': 'black' })
-})
+    $(this).css({ 'background-color': 'white', 'color': 'black' });
+});
 $('.add_btn').on('mouseover', function () {
-    $(this).css({ 'background-color': 'black', 'color': 'white' })
-})
+    $(this).css({ 'background-color': '#Fc3', 'color': 'black' });
+});
 $('.add_btn').on('mouseout', function () {
-    $(this).css({ 'background-color': 'white', 'color': 'black' })
-})
+    $(this).css({ 'background-color': 'white', 'color': 'black' });
+});
 let add_whiteboard = () => {
     let wb_id = Date.now();
     $('.whiteboard').append(`
@@ -71,46 +89,46 @@ let add_whiteboard = () => {
             <div class='wb_back'></div>
             <input class='wb_title' style='pointer-events: auto;'><img class='edit' src='./img/save.png'></input>
             <div class='close_btn'>X</div>
-        </div>`)
+        </div>`);
 
     /*---Style---*/
     $('.whiteboard').on('mouseover', '.wb_block', function () {
         $(this).children('.close_btn').show();
-        $(this).children('.edit').css({ 'visibility': 'visible' })
-    })
+        $(this).children('.edit').css({ 'visibility': 'visible' });
+    });
     $('.whiteboard').on('mouseout', '.wb_block', function () {
         $(this).children('.close_btn').hide();
-    })
-}
+    });
+};
 
 /*---Edit WB name---*/
 $('.whiteboard').on('click', '.edit', function () {
-    $(this).siblings('.wb_title').css({ 'pointer-events': 'auto', 'color': 'darkgray' })
-    $(this).attr('src', './img/save.png')
+    $(this).siblings('.wb_title').css({ 'pointer-events': 'auto', 'color': 'darkgray' });
+    $(this).attr('src', './img/save.png');
     $(this).addClass('save');
-})
+});
 /*---Save WB---*/
 $('.whiteboard').on('click', 'img[src="./img/save.png"]', function (e) {
-    $(this).siblings('.wb_title').css({ 'pointer-events': 'none', 'color': 'black' })
-    $(this).attr('src', './img/edit.png')
+    $(this).siblings('.wb_title').css({ 'pointer-events': 'none', 'color': 'black' });
+    $(this).attr('src', './img/edit.png');
     $(this).removeClass('save');
     createWhiteboard(e);
 
-})
+});
 
 function createWhiteboard(e) {
     // get wb title & wb_id & bookmark
     let title = $(e.target).siblings('.wb_title').val();
     let wb_id = $(e.target).parent('.wb_block').attr('id');
-    let bookmark = $(e.target).siblings('.wb_bookmark').data('bookmark') || $(e.target).data('bookmark')
+    let bookmark = $(e.target).siblings('.wb_bookmark').data('bookmark') || $(e.target).data('bookmark');
     if (bookmark == undefined) {
-        bookmark = 'null'
+        bookmark = 'null';
     }
 
     if (title.includes("'")) {
-        alertMessage("Cannot input symbols including:\" and \'", 'danger')
+        alertMessage("Cannot input symbols including:\" and \'", 'danger');
     } else {
-        fetch('/api/1.0/dashboard/createWhiteboard', {
+        fetch('/api/1.0/dashboard', {
             method: 'POST',
             headers: {
                 'content-type': 'application/json',
@@ -128,7 +146,7 @@ function createWhiteboard(e) {
                 } else {
                     alertMessage(error, 'danger');
                 }
-            })
+            });
     }
 
 }
@@ -139,48 +157,39 @@ $('.whiteboard').on('click', '.close_btn', function (e) {
     let title = $(e.target).siblings('.wb_title').val();
     let wb_id = $(e.target).parent('.wb_block').attr('id');
 
-    // (WIP)ask if delete
-    // Swal.fire({
-    //     title: 'Delete whiteboard?',
-    //     text: "You won't be able to revert this!",
-    //     icon: 'warning',
-    //     showCancelButton: true,
-    //     confirmButtonColor: '#3085d6',
-    //     cancelButtonColor: '#d33',
-    //     confirmButtonText: 'Yes, delete it!'
-    // }).then((result) => {
-    //     if (result.value) {
-    //         Swal.fire(
-    //             'Deleted!',
-    //             'Whiteboard has been deleted.',
-    //             'success'
-    //         )
-    //     }
-    // })
-    // alertMessage('Delete whiteboard?', 'info');
+    // Ask if delete
+    Swal.fire({
+        title: 'Delete whiteboard?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+        if (result.value) {
+            Swal.fire(
+                'Deleted!',
+                'Whiteboard has been deleted.',
+                'success'
+            );
+            fetch('api/1.0/dashboard', {
+                method: 'DELETE',
+                headers: {
+                    'content-type': 'application/json',
+                    'authorization': access_token
+                },
+                body: JSON.stringify({
+                    wb_id, title
+                })
+            });
 
-    fetch('api/1.0/dashboard/deleteWhiteboard', {
-        method: 'DELETE',
-        headers: {
-            'content-type': 'application/json',
-            'authorization': access_token
-        },
-        body: JSON.stringify({
-            wb_id, title
-        })
-    })
-        .then((res) => res.json())
-        .then((data) => {
-            let { message, error } = data;
-            if (message) {
-                alertMessage(message, 'success');
-            } else {
-                alertMessage(error, 'danger');
-            }
-        })
+            // delete whiteboard effect
+            e.target.closest('.wb_block').remove();
 
-    // delete whiteboard effect
-    e.target.closest('.wb_block').remove();
+        }
+    });
+
 });
 
 
@@ -192,57 +201,39 @@ $('section').on('click', '.wb_block', function (e) {
     let title = $(e.target).children('.wb_title').val();
     // redirect excluding input
     if (wb_id && title && $(e.target).hasClass('wb_block')) {
-        window.location.href = `workspace.html?wb_id=${wb_id}&title=${title}`
+        window.location.href = `workspace.html?wb_id=${wb_id}&title=${title}`;
     }
 
-})
+});
 
 /*---Bookmark---*/
 $('.whiteboard').on('click', '.wb_bookmark', function (e) {
     $(this).data('bookmark', 'bookmarked');
     $(this).addClass('bookmarked');
     createWhiteboard(e);
-})
+});
 $('.whiteboard').on('click', '.bookmarked', function (e) {
     $(this).removeData('bookmark');
     $(this).removeClass('bookmarked');
     createWhiteboard(e);
-})
+});
 
-// token verification
-if (access_token && username) {
-    $('.logout').html('LOGOUT');
-    $('.greeting').html(`Hello, ${username}`);
-    $('.greeting').css({ 'background-color': 'lightgray', 'color': 'black', 'border-radius': '2px', 'padding': '1px 5px 8px 5px', 'font-weight': 'bold' })
-} else {
-    $('.logout').html('LOGIN');
-    alert('Please Login, my friend!')
-    location.href = '/login.html'
-}
 
 // Logout
-$('.logout').click(function () {
+$('.logout').on('click', function () {
     if (access_token) {
         localStorage.removeItem('access_token');
         localStorage.removeItem('username');
-        location.href = '/logout.html';
+        localStorage.removeItem('user_id');
+        localStorage.removeItem('user_color');
+        alertMessage('You\'re logged out, have a great day!', 'success');
+        location.href = '/';
     } else {
-        location.href = '/login.html'
+        location.href = '/';
     }
-})
+});
 
 
-// Function: set multiple attributes
-function setAttributes(el, options) {
-    Object.keys(options).forEach(function (attr) {
-        el.setAttribute(attr, options[attr]);
-    })
-}
-
-
-/*---(WIP) Cursor Style---*/
-$('body').awesomeCursor('pencil',
-    { color: 'black' });
